+++ nextjs/src/lib/strapi.ts
@@ -126,7 +126,7 @@ export function getEventPage(locale: Locale, slug: string) {
 export function getEventPageCards(locale: Locale) {
   validateLocale(locale)
   return strapi(
-    `event-pages/?locale=${normalizeLocale(locale)}&populate=card_image&populate=hero.background_image&filters[is_past_event][$eq]=false&status=published`,
+    `event-pages/?locale=${normalizeLocale(locale)}&populate=card_image&populate=hero.background_image&filters[is_past_event][$eq]=false&status=published&sort=date_event:asc`,
   )
 }